Removed some warnings.  Removed old gif file.
Former-commit-id: 8c7ea6ff9e6173d1d60b8d12ae1067699da23e62

+++ mrViewer/src/standalone/mrvCommandLine.cpp
@@ -457,7 +457,7 @@ void parse_command_line( const int argc, char** argv,
                   mrv::get_sequence_limits( start, end, fileroot );
                }
 
-               if ( e - i <= files.size() - normalFiles )
+               if ( (size_t)(e - i) <= files.size() - normalFiles )
                {
                   // Add audio file to last stereo fileroot
                   if ( ai != ae )
